Try setting data type

+++ nemo_aligner/algorithms/reinforce_ifeval.py
@@ -263,7 +263,7 @@ class ReinforceIFEvalTrainer:
                         response = self.model.tokenizer.ids_to_text(rollout_batch["response_tokens"][i, rollout_batch["prompt_lengths"][i]:rollout_batch["response_lengths"][i]].tolist())
                         rewards.append(self.ifeval_rewards(prompt, response, args_duplicated[i]))
 
-                    rewards = torch.tensor(rewards, device=rollout_batch["logprobs"].device).unsqueeze(-1)
+                    rewards = torch.tensor(rewards, device=rollout_batch["logprobs"].device, dtype=torch.float32).unsqueeze(-1)
                     init_policy_logprobs = self.model.get_init_policy_logprobs([rollout_batch])[0]
 
                     if "rewards" in current_batch:
@@ -309,7 +309,7 @@ class ReinforceIFEvalTrainer:
                     response = self.model.tokenizer.ids_to_text(rollout_batch["response_tokens"][i, rollout_batch["prompt_lengths"][i]:rollout_batch["response_lengths"][i]].tolist())
                     rewards.append(self.ifeval_rewards(prompt, response, inference_batch["args"][i]))
 
-                rewards = torch.tensor(rewards, device=rollout_batch["logprobs"].device).unsqueeze(-1)
+                rewards = torch.tensor(rewards, device=rollout_batch["logprobs"].device, dtype=torch.float32).unsqueeze(-1)
                 rollout_batch["rewards"] = rewards
                 rollout_batches.append(rollout_batch)
